split and document serizliers
